add axis symbols
